Release 3.0 rc3.0: release candidate 3.0
The main changes are:
  - The functions: compressOccupancyMap and compressOccupancyMapInfo have been merged
  - Update encoding parameters
     - Cfg files
     - Deleting processes that force the value of the input parameters.
  - Source code cleaning

+++ source/lib/PccLibCommon/source/PCCPointSet.cpp
@@ -125,7 +125,7 @@ bool PCCPointSet3::write( const std::string &fileName, const bool asAscii ) {
         fout << " " << static_cast<int>(getReflectance(i));
       }
       if( PCC_SAVE_POINT_TYPE ) {
-        fout << " " << types_[ i ];
+        fout << " " <<  static_cast<int>( types_[ i ] );
       }
       fout << std::endl;
     }
